% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nls_models.R
\name{fun_nonrec}
\alias{fun_nonrec}
\title{functions for different light response models}
\usage{
fun_nonrec(x, Am, alpha, Rd, theta)
}
\arguments{
\item{x}{is the PAR intensity of the leaf}

\item{Am}{is the gross photosynthetic rate}

\item{alpha}{is  quantum yield at low light intensity}

\item{Rd}{dark respiration rate}

\item{theta}{is the convexity}

\item{Ic}{light compensation point}

\item{b}{an adjusting factor}
}
\description{
fun_nonrec for nonrectangular hyperbola–based models
fun_rec for hyperbolic tangent models and
fun_exp for exponential based models
}
\details{
aims to help us to confirm the range of the start values for nls analysis
base on ggplot2 geom_smooth, as we just need to sure the estimate values are 
as close to the measured values as possible, then the value are the start values
for nls.
}
\examples{
\dontrun{
# none, internal use
}
}
