% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rec.R
\name{plot_rec64}
\alias{plot_rec64}
\title{plot for hyperbolic tangent models models}
\usage{
plot_rec64(df, Am = max(df$Photo) - min(df$Photo), Rd = min(df$Photo),
  alpha = 0.01)
}
\arguments{
\item{Am}{is the gross photosynthetic rate}

\item{Rd}{dark respiration rate}

\item{alpha}{is  quantum yield at low light intensity}

\item{x}{is the PAR intensity of the leaf}
}
\description{
plots of different parameter values in one plot, to determin the range
of start values, plot_rec64 for LI-6400 data, and plot_rec68 for LI-6800 data.
}
\details{
use ggplot2 geom_smooth to fit the calculate data, generally the fit method
is loess, but it does not matter, as we just want to see these data overlap with the 
measured ones as much as possible.
}
\examples{
\dontrun{
plot_rec64(lrc64, Am = 23)
plot_rec68(lrc68, Am = 15, alpha = 0.01)
}
}
